flatten grpc editor
